Add Sender support
